Rename lib to be more clear about it's purpose
